requirements.txt
Flask==2.0.2
pandas==1.3.4
python-docx==0.8.11
gunicorn==20.1.0
openpyxl==3.0.9  # Add this line
